% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABF.R
\name{ABF}
\alias{ABF}
\title{Fine-map with ABF}
\source{
\href{https://www.nature.com/articles/ng.2435}{
JB Maller et al., Bayesian refinement of association signals for
 14 loci in 3 common diseases. Nature Genetics. 44, 1294–1301 (2012).}

\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1950810/}{
J Wakefield, A bayesian measure of the probability of false discovery 
in genetic epidemiology studies. American Journal of Human Genetics. 
81, 208–227 (2007).}
}
\usage{
ABF(
  dat,
  PP_threshold = 0.95,
  sample_size = NULL,
  sdY = NULL,
  case_control = TRUE,
  verbose = TRUE
)
}
\description{
Conduct statistical fine-mapping with approximate Bayes factor (ABF).
}
\examples{
dat <- echodata::LRRK2
dat2 <- echofinemap::ABF(dat=dat) 
}
