% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_args}
\alias{check_args}
\title{Check arguments}
\source{
\code{
finemap_args <- list(SUSIE=list("typo1"=1,rescale_priors=TRUE),
                      FINEMAP=list("FINEMAP_path"=NULL,"model"=NULL))
fma <- check_args(finemap_args=finemap_args, finemap_method="SUSIE")
}
}
\usage{
check_args(finemap_args, finemap_method, max_values = 1, verbose = TRUE)
}
\arguments{
\item{finemap_args}{A named nested list containing additional arguments 
for each fine-mapping method. e.g.
\code{finemap_args = list(FINEMAP=list(), PAINTOR=list(method=""))}}

\item{max_values}{Max number of values to return for any argument.}

\item{verbose}{Print messages.}
}
\value{
Named list of arguments for a specific \code{finemap_method}.
}
\description{
Check a list of arguments for a given function and:
\itemize{
\item{Remove any that do not match available arguments for that method.}
\item{Set default values for any arguments that are NULL or not provided.}
}
}
\keyword{internal}
