% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_necessary_cols.R
\name{check_necessary_cols}
\alias{check_necessary_cols}
\title{Check for necessary columns}
\usage{
check_necessary_cols(
  dat,
  finemap_methods,
  sample_size = NULL,
  dataset_type = "GWAS",
  verbose = TRUE
)
}
\description{
Check for necessary columns
}
\examples{
dat <- echodata::BST1;
finemap_methods <- c("ABF","FINEMAP","SUSIE","POLYFUN_SUSIE")
finemap_methods <- check_necessary_cols(dat=dat, 
                                        finemap_methods=finemap_methods)
}
