% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COJO_find_executable.R
\name{COJO_find_executable}
\alias{COJO_find_executable}
\title{COJO: find executable}
\usage{
COJO_find_executable(gcta_path = NULL, verbose = TRUE)
}
\value{
Executable path.
}
\description{
Find the path to the COJO executable.
}
\keyword{internal}
