% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COJO_get_stepwise_results.R
\name{COJO_get_stepwise_results}
\alias{COJO_get_stepwise_results}
\title{Gather stepwise conditional results}
\usage{
COJO_get_stepwise_results(
  cojo_dir = NULL,
  prefix = "cojo",
  jma_cojo_path = file.path(cojo_dir, paste0(prefix, ".jma.cojo")),
  verbose = TRUE
)
}
\arguments{
\item{prefix}{Prefix to use for file names.}

\item{verbose}{Print messages.}
}
\value{
independent_snps
}
\description{
Gather and preprocess the results of the \emph{GCTA-COJO} 
conditional stepwise procedure.
\href{https://yanglab.westlake.edu.cn/software/gcta}{
\strong{GTCA-COJO Documentation}:}\cr
"Perform a stepwise model selection procedure to select 
independently associated SNPs. Results will be saved in a 
\emph{.jma} file with additional file \emph{.jma.ldr} showing the LD 
correlations between the SNPs."
}
\concept{COJO
\url{https://www.nature.com/articles/ng.2213}
\url{https://www.cell.com/ajhg/fulltext/S0002-9297(10)00598-7}
\url{https://cnsgenomics.com/software/gcta/#Overview}}
\keyword{internal}
