% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COJO_process_results.R
\name{COJO_process_results}
\alias{COJO_process_results}
\title{Gather all \emph{GCTA-COJO} results}
\usage{
COJO_process_results(
  dat,
  paths,
  credset_thresh = 0.95,
  freq_cutoff = 0.1,
  verbose = TRUE
)
}
\description{
Gather all \emph{GCTA-COJO} results
}
\seealso{
Other COJO: 
\code{\link{COJO_locus_subdir}()},
\code{\link{COJO_run}()},
\code{\link{COJO}()}
}
\concept{COJO}
\keyword{internal}
