% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_finemap_cols.R
\name{drop_finemap_cols}
\alias{drop_finemap_cols}
\title{Drop fine-mapping columns}
\usage{
drop_finemap_cols(
  dat,
  finemap_methods = NULL,
  summary_cols = c("Support", "Consensus_SNP", "mean\\\\.PP"),
  verbose = TRUE
)
}
\arguments{
\item{dat}{Fine-mapping results data.}

\item{summary_cols}{Multi-tool summary columns to drop.}

\item{verbose}{Print messages.}
}
\description{
Drop all method-specific and/or multi-tool summary columns from data.
This is helpful when ensuring duplicate columns aren't created when 
re-running fine-mapping with the same method.
}
\examples{
dat2 <- echofinemap::drop_finemap_cols(dat=dat)
}
