% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINEMAP_construct_data.R
\name{FINEMAP_construct_data}
\alias{FINEMAP_construct_data}
\title{Prepare input files for \code{FINEMAP}}
\source{
\url{http://www.christianbenner.com}

\code{
locus_dir <- echodata::locus_dir;
dat <- echodata::BST1; 
LD_matrix <- echodata::BST1_LD_matrix
dir.create(file.path(locus_dir,"FINEMAP"), 
           showWarnings = FALSE, recursive = TRUE)
out <- echoLD::subset_common_snps(LD_matrix=LD_matrix, dat=dat)
LD_matrix <- out$LD
dat <- out$DT
dat_paths <- echofinemap:::FINEMAP_construct_data(dat=dat, 
                                                  locus_dir=locus_dir, 
                                                  LD_matrix=LD_matrix)
}
}
\usage{
FINEMAP_construct_data(dat, locus_dir, LD_matrix, nThread = 1, verbose = TRUE)
}
\description{
Creates and saves 1) the summary stats file, and 2) the LD matrix.
"Columns beta and se are required for fine-mapping.
Column maf is needed to output posterior effect size estimates on the
allelic scale. All other columns are not required for computations and
can be specified arbitrarily."
}
\seealso{
Other FINEMAP: 
\code{\link{FINEMAP_construct_master}()},
\code{\link{FINEMAP_find_executable}()},
\code{\link{FINEMAP_process_results}()},
\code{\link{FINEMAP}()}
}
\concept{FINEMAP}
\keyword{internal}
