% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINEMAP_find_executable.R
\name{FINEMAP_find_executable}
\alias{FINEMAP_find_executable}
\title{Find FINEMAP executable}
\source{
\url{http://www.christianbenner.com}
}
\usage{
FINEMAP_find_executable(
  FINEMAP_path = NULL,
  OS = NULL,
  save_dir = file.path(tools::R_user_dir(package = "echofinemap", which = "cache"),
    "FINEMAP"),
  version = "1.4",
  verbose = TRUE
)
}
\description{
Retrieve location of \code{FINEMAP} executable.
}
\examples{
FINEMAP_path <- echofinemap:::FINEMAP_find_executable(version="1.3.1")
}
\seealso{
Other FINEMAP: 
\code{\link{FINEMAP_construct_data}()},
\code{\link{FINEMAP_construct_master}()},
\code{\link{FINEMAP_process_results}()},
\code{\link{FINEMAP}()}
}
\concept{FINEMAP}
\keyword{internal}
