% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap_handler.R
\name{finemap_handler}
\alias{finemap_handler}
\title{Fine-map handler}
\usage{
finemap_handler(
  locus_dir,
  fullSS_path,
  finemap_methods = c("SUSIE", "FINEMAP"),
  finemap_args = NULL,
  dat,
  dataset_type = "GWAS",
  force_new_finemap = TRUE,
  LD_reference = NULL,
  LD_matrix = NULL,
  n_causal = 5,
  sample_size = NULL,
  conditioned_snps = NULL,
  PAINTOR_QTL_datasets = NULL,
  PP_threshold = 0.95,
  consensus_threshold = 2,
  case_control = TRUE,
  conda_env = "echoR",
  nThread = 1,
  verbose = TRUE
)
}
\description{
Handle fine-mapping across multiple tools.
}
\seealso{
Other finemapping functions: 
\code{\link{multi_finemap}()}
}
\concept{finemapping functions}
