% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINEMAP_process_results.R
\name{FINEMAP_process_results}
\alias{FINEMAP_process_results}
\title{FINEMAP: process results}
\source{
\url{http://www.christianbenner.com}
}
\usage{
FINEMAP_process_results(
  dat,
  locus_dir,
  credset_thresh = 0.95,
  pvalue_thresh = 0.05,
  finemap_version = "1.4",
  results_file = ".cred",
  nThread = 1,
  sort_by_CS = TRUE,
  verbose = TRUE
)
}
\description{
Post-processing of \code{FINEMAP} results.
}
\examples{
\dontrun{
locus_dir <- file.path(tempdir(), echodata::locus_dir)
dat <- echodata::BST1;
dat2 <-FINEMAP_process_results(dat=dat, locus_dir=locus_dir)
}
}
\seealso{
Other FINEMAP: 
\code{\link{FINEMAP_construct_data}()},
\code{\link{FINEMAP_construct_master}()},
\code{\link{FINEMAP_find_executable}()},
\code{\link{FINEMAP}()}
}
\concept{FINEMAP}
\keyword{internal}
