% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pheno_variance.R
\name{get_pheno_variance}
\alias{get_pheno_variance}
\title{Compute phenotype variance}
\usage{
get_pheno_variance(dat, dataset_type, var_name = "Expression", verbose = TRUE)
}
\description{
Compute phenotype variance (\code{var_y} in \pkg{susieR}).
}
\keyword{internal}
