% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_submodule_git.R
\name{install_submodule_git}
\alias{install_submodule_git}
\title{Install a submodule}
\source{
\href{https://github.com/r-lib/devtools/pull/751#issuecomment-215540807}{
GitHub Issue}
}
\usage{
install_submodule_git(x, ...)
}
\description{
Install a submodule from a remote GitHub repository.
}
\keyword{internal}
