% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfm.R
\name{lfm}
\alias{lfm}
\title{List fine-mapping methods}
\usage{
lfm(finemap_methods = NULL, verbose = TRUE)
}
\arguments{
\item{finemap_methods}{Fine-mapping methods to run. 
See \link[echofinemap]{lfm} for a list of all fine-mapping methods currently
available.}

\item{verbose}{Print messages.}
}
\description{
List all fine-mapping methods currently available in \pkg{echofinemap}.
}
\examples{
finemap_methods <- echofinemap::lfm()
}
