% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_finemap.R
\name{multi_finemap}
\alias{multi_finemap}
\title{Fine-map using multiple fine-mapping tools}
\usage{
multi_finemap(
  locus_dir,
  fullSS_path,
  finemap_method_list,
  finemap_args = NULL,
  dat,
  dataset_type,
  LD_matrix = NULL,
  n_causal = 5,
  sample_size = NULL,
  PAINTOR_QTL_datasets = NULL,
  PP_threshold = 0.95,
  case_control = TRUE,
  verbose = TRUE,
  nThread = 1,
  conda_env = "echoR"
)
}
\description{
Fine-mapping will be repeated on the same locus using each of the tools 
in \code{finemap_method_list}.
Then, all results will be merged into the locus-specific multi-finemap file,
along with the original per-SNP GWAS/QTL summary statistics.
Each tools will have the following columns:
\describe{
 \item{<tool>.PP}{The posterior probability (PP) of a SNP being causal for
  the trait. Though this is a generalization and the exact meaning of PP
   will differ by tools (e.g. Posterior Inclusion Probability for SUSIE).}
 \item{<tool>.CS}{Which credible set the SNP is part of (within a locus). 
 If \code{=0}, then the SNP was not part of any credible set. 
 Some tools only produce one credible set per locus.}
}
}
\examples{
\dontrun{
BST1 <- echodata::BST1; data("BST1_LD_matrix");
dat <- BST1
finemap_method_list <- c("ABF","SUSIE")
}
}
\concept{finemapping functions}
\keyword{internal}
