% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR}
\alias{PAINTOR}
\title{Run full PAINTOR pipeline}
\source{
\href{https://github.com/gkichaev/PAINTOR_V3.0}{GitHub}
\href{https://github.com/gkichaev/PAINTOR_V3.0/wiki/2a.-Computing-1000-genomes-LD}{LD Tutorial}
}
\usage{
PAINTOR(
  dat,
  annot = NULL,
  dat_prefixes = NULL,
  annot_prefixes = NULL,
  zscore_col = "ZSCORE",
  tstat_col = "tstat",
  LD_matrix,
  locus_dir,
  fullSS_path,
  storage_dir = tempdir(),
  paintor_path = NULL,
  n_causal = 5,
  XGR_dataset = NA,
  ROADMAP_search = NA,
  chromatin_state = "TssA",
  use_annotations = FALSE,
  credset_thresh = 0.95,
  consensus_thresh = 2,
  multi_finemap_col_name = "PAINTOR",
  trim_gene_limits = FALSE,
  force_new_subset = FALSE,
  dat_populations = "EUR",
  LD_reference = "1KGphase3",
  force_new_LD = FALSE,
  force_reinstall = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{dat}{A \link[data.table]{data.table} of GWAS/QTL/TWAS 
summary statistics. Alternatively, for multi-trait/multi-ancestry
 fine-mapping you can instead provide a named list of 
 \link[data.table]{data.table}s with overlapping genomic coordinates 
 (e.g. \code{list(GWAS1=dat1, GWAS2=dat2)}).}
}
\description{
Run full PAINTOR pipeline
}
\examples{
dat <- echodata::BST1
LD_matrix <- echodata::BST1_LD_matrix
fullSS_path <- echodata::example_fullSS()
locus_dir <- file.path(tempdir(),echodata::locus_dir)
dat2 <- PAINTOR(dat = list("Nalls2019"=dat),
                locus_dir = locus_dir)
}
