% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_create_locus_file.R
\name{PAINTOR_create_locus_file}
\alias{PAINTOR_create_locus_file}
\title{Create QTL locus file for PAINTOR}
\usage{
PAINTOR_create_locus_file(
  dat_merged,
  LD_ls,
  locus_dir,
  PT_results_path,
  NA_method = c("drop", "fill")
)
}
\description{
Create QTL locus file for PAINTOR
}
\keyword{internal}
