% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_datatype_handler.R
\name{PAINTOR_datatype_handler}
\alias{PAINTOR_datatype_handler}
\title{Determine data types (GWAS, QTL)}
\usage{
PAINTOR_datatype_handler(
  locus_dir,
  dat_prefixes = NULL,
  annot_prefixes = NULL,
  verbose = TRUE
)
}
\description{
Determine data types (GWAS, QTL)
}
\keyword{internal}
