% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_gather_transethnic_LD.R
\name{PAINTOR_gather_transethnic_LD}
\alias{PAINTOR_gather_transethnic_LD}
\title{Prepare transethnic PAINTOR results}
\usage{
PAINTOR_gather_transethnic_LD(
  merged_DT,
  locus,
  conditions = c("Nalls23andMe_2019", "MESA_AFA", "MESA_CAU", "MESA_HIS")
)
}
\description{
Prepare transethnic PAINTOR results
}
\concept{PAINTOR}
\keyword{internal}
