% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_import_qtl_dt.R
\name{PAINTOR_import_qtl_dat}
\alias{PAINTOR_import_qtl_dat}
\title{Import QTL data for PAINTOR}
\usage{
PAINTOR_import_qtl_dat(
  QTL_datasets,
  locus_dir,
  trim_gene_limits = FALSE,
  force_new_subset = FALSE,
  metric = "t_stat",
  verbose = TRUE
)
}
\description{
@keywords internal
}
