% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_install.R
\name{PAINTOR_install}
\alias{PAINTOR_install}
\title{Install PAINTOR via command line}
\source{
\href{https://github.com/gkichaev/PAINTOR_V3.0/issues/49}{
Installation issues with gcc versions}
}
\usage{
PAINTOR_install(paintor_path = NULL, force_reinstall = FALSE, verbose = TRUE)
}
\description{
Compile PAINTOR from source using the bundled install script.
}
\keyword{internal}
