% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_locusName_handler.R
\name{PAINTOR_locusName_handler}
\alias{PAINTOR_locusName_handler}
\title{Construct locus name for PAINTOR}
\usage{
PAINTOR_locusName_handler(
  locus_name = NULL,
  locus,
  GWAS_datasets = NULL,
  QTL_datasets = NULL
)
}
\description{
Construct locus name for PAINTOR
}
\keyword{internal}
