% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_merge_datasets.R
\name{PAINTOR_merge_datasets}
\alias{PAINTOR_merge_datasets}
\title{Merge datasets}
\usage{
PAINTOR_merge_datasets(dat_ls, NA_method = c("drop", "fill"))
}
\description{
Merge a named list of GWAS/QTL/TWAS datasets.
}
\keyword{internal}
