% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_merge_results.R
\name{PAINTOR_merge_results}
\alias{PAINTOR_merge_results}
\title{Merge PAINTOR results}
\usage{
PAINTOR_merge_results(
  dat,
  PAINTOR_results,
  credset_thresh = 0.5,
  multi_finemap_col_name = "PAINTOR"
)
}
\description{
@keywords internal
}
