% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_prepare_LD.R
\name{PAINTOR_prepare_LD}
\alias{PAINTOR_prepare_LD}
\title{Prepare LD file for PAINTOR}
\usage{
PAINTOR_prepare_LD(
  LD_matrix,
  dat_merged,
  locus_dir,
  PT_results_path,
  verbose = TRUE
)
}
\description{
\strong{VERY IMPORTANT!}\cr
"Particular care must be taken when computing LD from a reference panel 
such the 1000 genomes.
It is imperative that all the reference and alternate alleles for SNPs
from which the Z-scores were computed match the reference and 
alternate alleles of the reference panel.
The output of PAINTOR will not be correct if there are mismatches of
this type in the data.
Please see wiki section 2a for instructions on how to use the LD 
utility provided with the software."
}
\keyword{internal}
