% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_prepare_ld_multiancestry.R
\name{PAINTOR_prepare_ld_multiancestry}
\alias{PAINTOR_prepare_ld_multiancestry}
\title{Prepare multi-ancestry LD files for PAINTOR}
\usage{
PAINTOR_prepare_ld_multiancestry(
  dat_merged,
  LD_ls,
  locus_dir,
  dat_populations,
  PT_results_path,
  LD_reference = "1KGphase3",
  force_new_LD = FALSE,
  shared_snps_only = TRUE,
  fillNA = 0,
  nThread = 1,
  verbose = TRUE
)
}
\description{
Prepare multi-ancestry LD files for PAINTOR
}
\keyword{internal}
