% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_process_results.R
\name{PAINTOR_process_results}
\alias{PAINTOR_process_results}
\title{Process PAINTOR results}
\usage{
PAINTOR_process_results(dat_merged, res_paths, credset_thresh, verbose = TRUE)
}
\description{
@keywords internal
 @importFrom data.table rbindlist fread
 @importFrom dplyr rename
 @importFrom stats setNames
}
