% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_run.R
\name{PAINTOR_run}
\alias{PAINTOR_run}
\title{Run PAINTOR}
\usage{
PAINTOR_run(
  paintor_path = NULL,
  PT_results_path,
  inputFile_path,
  ld_paths,
  zscore_cols,
  max_causal,
  Gname = "Enrichment.Estimates.txt",
  RESname = "results.txt",
  ANname = "annotations.txt",
  method = c("mcmc", "enumerate"),
  set_seed = 2019,
  auto_restart = FALSE,
  verbose = TRUE
)
}
\description{
@keywords internal
}
