% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_sample_data.R
\name{PAINTOR_sample_data}
\alias{PAINTOR_sample_data}
\title{Find PAINTOR sample data}
\usage{
PAINTOR_sample_data(
  annot_dir = system.file("tools", "PAINTOR_V3.0", "SampleData", package = "echofinemap"),
  pattern = "\\\\.annotations",
  recursive = TRUE,
  read = FALSE,
  verbose = TRUE
)
}
\description{
Find PAINTOR sample data
}
\keyword{internal}
