% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_survey_annotation.R
\name{PAINTOR_survey_annotation}
\alias{PAINTOR_survey_annotation}
\title{Report the number of annotations > 0}
\usage{
PAINTOR_survey_annotation(PT_results_path, locus_name = "Locus1")
}
\description{
@keywords internal
}
