% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_check_version.R
\name{POLYFUN_check_version}
\alias{POLYFUN_check_version}
\title{POLYFUN: check version}
\usage{
POLYFUN_check_version(polyfun = NULL, conda_env = "echoR_mini", verbose = TRUE)
}
\arguments{
\item{polyfun}{[Optional] Path to PolyFun executable.}

\item{conda_env}{Conda environment name.}

\item{verbose}{Print messages.}
}
\description{
Check PolyFun version installed.
}
\keyword{internal}
