% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN.R
\name{POLYFUN_compute_priors}
\alias{POLYFUN_compute_priors}
\title{Recompute SNP-wise priors from summary stats}
\usage{
POLYFUN_compute_priors(
  polyfun = NULL,
  locus_dir,
  fullSS_path,
  sample_size = NULL,
  min_INFO = 0,
  min_MAF = 0.001,
  annotations_path = NULL,
  weights_path = NULL,
  prefix = "PD_GWAS",
  chrom = "all",
  compute_ldscores = FALSE,
  allow_missing_SNPs = TRUE,
  ref_prefix = NULL,
  remove_tmps = TRUE,
  conda_env = "echoR"
)
}
\description{
Recompute SNP-wise priors from summary stats
}
\examples{
\dontrun{
locus_dir <- echodata::locus_dir;
fullSS_path <- example_fullSS(fullSS_path="~/Desktop/Nalls23andMe_2019.fullSS_subset.tsv")
munged_path <- "results/GWAS/Nalls23andMe_2019/_genome_wide/PolyFun/nallsEtAl2019_allSamples_allVariants.mod.munged.parquet"
LDSC.files <- POLYFUN_compute_priors(locus_dir=locus_dir, fullSS_path=fullSS_path, conda_env="echoR")
}
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_SUSIE}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_polyfun_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_get_precomputed_priors}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()},
\code{\link{read_parquet}()}
}
\concept{polyfun}
\keyword{internal}
