% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_find_polyfun_folder.R
\name{POLYFUN_find_polyfun_folder}
\alias{POLYFUN_find_polyfun_folder}
\title{Find PolyFun folder}
\source{
\code{
polyfun_path <- POLYFUN_find_polyfun_folder()
}
}
\usage{
POLYFUN_find_polyfun_folder(polyfun_path = NULL)
}
\description{
Folder where PolyFun submodule is stored.
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_import_priors}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()},
\code{\link{POLYFUN}()}
}
\concept{polyfun}
\keyword{internal}
