% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_finemapper.R
\name{POLYFUN_finemapper}
\alias{POLYFUN_finemapper}
\title{Run \emph{PolyFun+SUSIE} fine-mapping pipeline}
\source{
https://www.nature.com/articles/s41588-020-00735-5
}
\usage{
POLYFUN_finemapper(
  polyfun = NULL,
  dat = NULL,
  npz_gz_LD = NULL,
  locus = NULL,
  sample_size = NULL,
  locus_dir,
  n_causal = 5,
  method = "susie",
  h2_path = NULL,
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\description{
Run \emph{PolyFun+SUSIE} fine-mapping pipeline
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_folder}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_import_priors}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()},
\code{\link{POLYFUN}()}
}
\concept{polyfun}
\keyword{internal}
