% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_gather_ldscores.R
\name{POLYFUN_gather_ldscores}
\alias{POLYFUN_gather_ldscores}
\title{Find and import LD score files}
\usage{
POLYFUN_gather_ldscores(output_prefix)
}
\description{
Find and import LD score files
}
\examples{
\dontrun{
output_prefix <- file.path(system.file("tools/polyfun/gold","",package = "echofinemap"),"testrun.22")
ldscore <- POLYFUN_gather_ldscores(output_prefix=output_prefix)
}
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_import_priors}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()},
\code{\link{POLYFUN}()}
}
\concept{polyfun}
\keyword{internal}
