% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_get_precomputed_priors.R
\name{POLYFUN_get_precomputed_priors}
\alias{POLYFUN_get_precomputed_priors}
\title{Extract pre-computed prior}
\source{
https://www.biorxiv.org/content/10.1101/807792v3
}
\usage{
POLYFUN_get_precomputed_priors(
  polyfun = NULL,
  locus_dir,
  dat = NULL,
  force_new_priors = TRUE,
  remove_tmps = FALSE,
  nThread = 1,
  conda_env = "echoR"
)
}
\description{
Extract SNP-wise prior probabilities pre-computed from many UK Biobank traits.
}
\examples{
\dontrun{

dat <- echodata::BST1; 
locus_dir <- echodata::locus_dir;
priors <- POLYFUN_get_precomputed_priors(locus_dir=locus_dir, dat=dat)
}
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_SUSIE}()},
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_polyfun_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()}
}
\concept{polyfun}
\keyword{internal}
