% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_h2_enrichment_SNPgroups_plot.R
\name{POLYFUN_h2_enrichment_SNPgroups_plot}
\alias{POLYFUN_h2_enrichment_SNPgroups_plot}
\title{Plot heritability (h2) enrichment}
\usage{
POLYFUN_h2_enrichment_SNPgroups_plot(
  RES,
  snp_groups = c("GWAS lead", "UCS", "Consensus (-PolyFun)", "Consensus"),
  comparisons_filter = function(x) {     if ("Consensus" \%in\% x)          return(x) },
  method = "wilcox.test",
  remove_outliers = TRUE,
  title = "S-LDSC heritability enrichment",
  xlabel = "SNP group",
  ylabel = bquote(~"h"^2 ~ "enrichment"),
  show_xtext = TRUE,
  show_padj = TRUE,
  show_signif = TRUE,
  vjust_signif = 0.5,
  show_plot = TRUE,
  save_path = FALSE
)
}
\description{
Plot heritability (h2) enrichment
}
\examples{
\dontrun{
root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping"
merged_dat <- merge_finemapping_results(dataset = file.path("Data/GWAS/Nalls23andMe_2019"), LD_reference = "UKB", minimum_support = 0)
RES <- POLYFUN_h2_enrichment_SNPgroups(merged_dat=merged_dat, out.path="/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide/PolyFun/output")

plot.h2 <- POLYFUN_h2_enrichment_SNPgroups_plot(RES = RES, show_plot = TRUE)
}
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_SUSIE}()},
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_polyfun_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_get_precomputed_priors}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()}
}
\concept{polyfun}
