% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_import_priors.R
\name{POLYFUN_import_priors}
\alias{POLYFUN_import_priors}
\title{Import pre-computed priors}
\source{
https://www.nature.com/articles/s41588-020-00735-5
}
\usage{
POLYFUN_import_priors(
  locus_dir,
  dat = NULL,
  polyfun = NULL,
  force_new_priors = TRUE,
  remove_tmps = FALSE,
  nThread = 1,
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\description{
Import SNP-wise prior probabilities pre-computed from many UK Biobank traits.
This function handles finding the intersection of SNPS that exist in the 
input GWAS summary stats \code{dat} and the pre-computed priors that come 
shipped with PolyFun. Then, it saves this subset as a new file for PolyFun
(or other fine-mapping tools) to use as input. 
Uses the \code{extract_snpvar.py} script from PolyFun.
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_polyfun_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()},
\code{\link{POLYFUN}()}
}
\concept{polyfun}
\keyword{internal}
