% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_initialize.R
\name{POLYFUN_initialize}
\alias{POLYFUN_initialize}
\title{Create output dir and import SNP data.frame}
\source{
\code{
BST1 <- echodata::BST1; locus_dir <- echodata::locus_dir;
dat <- POLYFUN_initialize(locus_dir=locus_dir, dat=BST1)
}
}
\usage{
POLYFUN_initialize(locus_dir, dat = NULL, nThread = 1)
}
\description{
Create output dir and import SNP data.frame
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_SUSIE}()},
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_polyfun_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_get_precomputed_priors}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()}
}
\concept{polyfun}
\keyword{internal}
