% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_plot.R
\name{POLYFUN_plot}
\alias{POLYFUN_plot}
\title{Plot PolyFun and other fine-mapping results}
\source{
https://www.nature.com/articles/s41588-020-00735-5
}
\usage{
POLYFUN_plot(
  dat,
  LD_matrix,
  locus = NULL,
  subtitle = "PolyFun Comparison",
  conditions = c("SUSIE", "POLYFUN_SUSIE", "FINEMAP", "PAINTOR", "PAINTOR_Fairfax")
)
}
\description{
Plot PolyFun and other fine-mapping results
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_SUSIE}()},
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_polyfun_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_import_priors}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()}
}
\concept{polyfun}
\keyword{internal}
