% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_prepare_snp_input.R
\name{POLYFUN_prepare_snp_input}
\alias{POLYFUN_prepare_snp_input}
\title{Prepare SNP input for PolyFun}
\usage{
POLYFUN_prepare_snp_input(PF.output.path, locus_dir, dat = NULL, nThread = 1)
}
\description{
PolyFun requires a space-delimited (gzipped or not) file with these columns:
\itemize{
\item{CHR}
\item{BP}
\item{A1}
\item{A2}
}
}
\examples{
BST1 <- echodata::BST1; locus_dir <- echodata::locus_dir;
dat <- BST1
PF.output.path <- file.path(locus_dir, "PolyFun")
POLYFUN_prepare_snp_input(PF.output.path=PF.output.path, locus_dir=locus_dir, dat=dat)
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_polyfun_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_import_priors}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_run_ldsc}()},
\code{\link{POLYFUN}()}
}
\concept{polyfun}
\keyword{internal}
