% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_read_priors.R
\name{POLYFUN_read_priors}
\alias{POLYFUN_read_priors}
\title{Read priors}
\usage{
POLYFUN_read_priors(snp_w_priors.file, nThread = 1, verbose = TRUE)
}
\description{
Read PolyFun prior probabilities.
 Handles PolyFun version differences in file formatting
  (see \href{https://github.com/RajLabMSSM/echolocatoR/issues/80}{here}).
}
\keyword{internal}
