% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_run_ldsc.R
\name{POLYFUN_run_ldsc}
\alias{POLYFUN_run_ldsc}
\title{Run a modified version of S-LDSC}
\source{
https://www.nature.com/articles/s41588-020-00735-5
}
\usage{
POLYFUN_run_ldsc(
  polyfun = NULL,
  output_dir = NULL,
  munged.path,
  min_INFO = 0.6,
  min_MAF = 0.05,
  annotations.path = file.path(polyfun, "example_data/annotations."),
  weights.path = file.path(polyfun, "example_data/weights."),
  prefix = "LDSC",
  chrom = "all",
  compute_ldscores = FALSE,
  allow_missing_SNPs = TRUE,
  munged_path =
    "/sc/arion/projects/pd-omics/tools/polyfun/Nalls23andMe_2019.sumstats_munged.parquet",
  ref.prefix = "/sc/arion/projects/pd-omics/data/1000_Genomes/Phase1/1000G.mac5eur.",
  freq.prefix = "/sc/arion/projects/pd-omics/tools/polyfun/1000G_frq/1000G.mac5eur.",
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\description{
Modifications to S-LDSC include L2-regularization.
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_polyfun_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_import_priors}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN}()}
}
\concept{polyfun}
\keyword{internal}
