% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_priors.R
\name{prepare_priors}
\alias{prepare_priors}
\title{Prepare priors}
\usage{
prepare_priors(
  dat,
  priors_col = NULL,
  snp_col = "SNP",
  rescale_priors = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Fine-mapping results data.}

\item{priors_col}{[Optional] Name of the a column in 
\code{dat} to extract SNP-wise prior probabilities from.}

\item{snp_col}{Name of the SNP column.}

\item{rescale_priors}{If prior probabilities are supplied,
rescale them from 0-1 (i.e. \code{rescaled_priors = priors / sum(priors)}).}

\item{verbose}{Print messages.}
}
\description{
Prepare values to be used as per-SNP priors probabilities in fine-mapping.
}
\keyword{internal}
