% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_filelist.R
\name{rbind_filelist}
\alias{rbind_filelist}
\title{Bind stored files}
\usage{
rbind_filelist(file.list, nThread = 1, verbose = TRUE)
}
\description{
Rapidly read a list of files from storage and concatenate them by rows.
}
\concept{general}
\keyword{internal}
