% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read parquet file as data.frame}
\usage{
read_parquet(
  parquet_path,
  conda_env = "echoR",
  method = "pandas",
  verbose = TRUE
)
}
\arguments{
\item{method}{Specify whether to use \code{\link{SparkR}} (R)
or \code{pandas} (Python + \code{reticulate}).}
}
\description{
Read parquet file as data.frame
}
\examples{
\dontrun{
parquet_path <- system.file("tools/polyfun/example_data/weights.10.l2.ldscore.parquet", package = "echolocatoR")
#### Using python (pandas) - default ####
dat <- read_parquet(parquet_path=parquet_path, method="pandas")

#### Using R (SparkR) ####
dat <- read_parquet(parquet_path=parquet_path, method="sparkR")
}
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_SUSIE}()},
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_polyfun_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_get_precomputed_priors}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()}
}
\concept{polyfun}
