% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dup_snps.R
\name{remove_dup_snps}
\alias{remove_dup_snps}
\title{Remove duplicate SNPs}
\usage{
remove_dup_snps(path, remove_dup = TRUE, verbose = TRUE)
}
\arguments{
\item{path}{Path to file.}

\item{remove_dup}{Remove duplicate SNPs from the file.}

\item{verbose}{Print messages.}
}
\description{
Remove any duplicate SNPs and rewrite the file to disk.
}
\keyword{internal}
