% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na_rows.R
\name{remove_na_rows}
\alias{remove_na_rows}
\title{Remove NA rows}
\usage{
remove_na_rows(
  dat,
  cols = c("Effect", "P", "StdErr", "SNP", "MAF"),
  verbose = TRUE
)
}
\arguments{
\item{dat}{Data}

\item{cols}{Column names.}

\item{verbose}{Print messages.}
}
\value{
Filtered data
}
\description{
Remove rows with NAs in one or more specific columns.
If the column is not present, it will not be used for filtering.
}
\keyword{internal}
