% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gcc.R
\name{setup_gcc}
\alias{setup_gcc}
\title{Setup gcc}
\usage{
setup_gcc(version = "latest", verbose = TRUE)
}
\arguments{
\item{version}{gcc version to use (e.g. 7,8,9,10).
If \code{"latest"}, the latest version of gcc 
currently installed on your machine will be used.}

\item{verbose}{Print messages.}
}
\value{
A data.frame listing all gcc versions, their paths, 
and which one is currently active.
}
\description{
Set up a symlink to override xcode's clang with a valid version of gcc.
}
\keyword{internal}
