% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUSIE.R
\name{SUSIE}
\alias{SUSIE}
\title{Fine-map with SUSIE}
\source{
\href{https://stephenslab.github.io/susieR/}{GitHub}
\href{https://rss.onlinelibrary.wiley.com/doi/full/10.1111/rssb.12388}{
Publication}
}
\usage{
SUSIE(
  dat,
  LD_matrix,
  dataset_type = "GWAS",
  max_causal = 5,
  sample_size = NULL,
  prior_weights = NULL,
  PP_threshold = 0.95,
  scaled_prior_variance = 0.001,
  estimate_residual_variance = FALSE,
  estimate_prior_variance = TRUE,
  residual_variance = NULL,
  max_iter = 100,
  estimate_prior_method = "optim",
  manual_var_y = FALSE,
  rescale_priors = TRUE,
  plot_track_fit = FALSE,
  return_all_CS = TRUE,
  file_prefix = file.path(tempdir(), "SUSIE"),
  var_name = "Expression",
  verbose = TRUE
)
}
\arguments{
\item{max_causal}{The maximum number of non-zero effects 
(and thus causal variants).}

\item{prior_weights}{A vector of length p, in which each entry
gives the prior probability that corresponding column of X has a
nonzero effect on the outcome, y.}

\item{scaled_prior_variance}{The prior variance, divided by
\code{var(y)} (or by \code{(1/(n-1))yty} for
\code{susie_suff_stat}); that is, the prior variance of each
non-zero element of b is \code{var(y) * scaled_prior_variance}. The
value provided should be either a scalar or a vector of length
\code{L}. If \code{estimate_prior_variance = TRUE}, this provides
initial estimates of the prior variances.}

\item{estimate_residual_variance}{If
\code{estimate_residual_variance = TRUE}, the residual variance is
estimated, using \code{residual_variance} as an initial value. If
\code{estimate_residual_variance = FALSE}, the residual variance is
fixed to the value supplied by \code{residual_variance}.}

\item{estimate_prior_variance}{If \code{estimate_prior_variance =
TRUE}, the prior variance is estimated (this is a separate
parameter for each of the L effects). If provided,
\code{scaled_prior_variance} is then used as an initial value for
the optimization. When \code{estimate_prior_variance = FALSE}, the
prior variance for each of the L effects is determined by the
value supplied to \code{scaled_prior_variance}.}

\item{residual_variance}{Variance of the residual. If
\code{estimate_residual_variance = TRUE}, this value provides the
initial estimate of the residual variance. By default, it is set to
\code{var(y)} in \code{susie} and \code{(1/(n-1))yty} in
\code{susie_suff_stat}.}

\item{max_iter}{Maximum number of IBSS iterations to perform.}

\item{estimate_prior_method}{The method used for estimating prior
variance. When \code{estimate_prior_method = "simple"} is used, the
likelihood at the specified prior variance is compared to the
likelihood at a variance of zero, and the setting with the larger
likelihood is retained.}

\item{rescale_priors}{If prior probabilities are supplied,
rescale them from 0-1 (i.e. \code{rescaled_priors = priors / sum(priors)}).}

\item{plot_track_fit}{Record each iteration and make a GIF of the
fine-mapping algorithm learning the causal variants.
\strong{WARNING!:} Making this plot can take a long time if there's
 many iterations.}

\item{file_prefix}{Prefix to path of output plot file. If not
specified, the plot, or plots, will be saved to a temporary
directory generated using \code{\link{tempdir}}.}

\item{verbose}{If \code{verbose = TRUE}, the algorithm's progress,
and a summary of the optimization settings, are printed to the
console.}
}
\description{
Sum of Single Effects (SuSiE): Iterative Bayesian Step-wise Selection.
}
\details{
\strong{Notes on convergence:}
\pkg{susieR} will often give the warning: 
\code{IBSS algorithm did not converge in 100 iterations!}.
This means the results might not necessarily be reliable.
There's several things you can try to avoid this:
\itemize{
\item{Make sure \code{susieR} is up-to-date: 
\code{ devtools::install_github("stephenslab/susieR@0.9.0")}}
\item{Increase \code{max_causal} (e.g. 5 => 10).}
\item{Increase \code{max_iter} (e.g. 100 => 1000), 
though this will take longer.}
\item{Decrease the locus window size, which will also speed up
 the algorithm but potentially miss causal variants far from the lead SNP.}
}
Changing \code{estimate_prior_method} does not seem to affect 
convergence warnings.

\strong{Notes on variance:}
\href{https://github.com/stephenslab/susieR/issues/90}{GitHub Issue}
If \code{estimate_residual_variance=TRUE} \emph{without} 
providing \code{var_y} \emph{and} \code{L>1}, \pkg{susieR} will throw error:
\code{Estimating residual variance failed: the estimated value is negative}
Running \pkg{susieR} with \code{var_y = var(b)} provides \emph{exactly}
 the same results.
}
\examples{
dat <- echodata::BST1
LD_matrix <- echodata::BST1_LD_matrix
dat2 <- echofinemap::SUSIE(dat=dat, LD_matrix=LD_matrix)
}
