% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUSIE_get_func.R
\name{SUSIE_get_func}
\alias{SUSIE_get_func}
\title{Get susieR fine-mapping function}
\usage{
SUSIE_get_func(verbose = TRUE)
}
\arguments{
\item{verbose}{Print messages.}
}
\description{
The authors of \pkg{susieR} merged \code{susie_ss} and \code{susie_bhat}
 into \code{susie_suff_stat} in 11/2019.
 They then changed \code{susie_ss} to \code{susie_rss}.
}
\keyword{internal}
