% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUSIE_run.R
\name{SUSIE_run}
\alias{SUSIE_run}
\title{Run \pkg{susieR} fine-mapping}
\usage{
SUSIE_run(
  dat,
  LD_matrix,
  sample_size,
  max_causal,
  scaled_prior_variance,
  estimate_prior_variance,
  residual_variance,
  max_iter,
  estimate_prior_method,
  estimate_residual_variance,
  prior_weights,
  PP_threshold,
  plot_track_fit,
  verbose = TRUE
)
}
\description{
Finds the proper function to call from \pkg{susieR}
 (which can depend on which version is installed) and conducts 
 statistical or functional fine-mapping using GWAS/QTL summary stats.
}
\keyword{internal}
