% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABF.R
\name{ABF}
\alias{ABF}
\title{Fine-map with ABF}
\source{
\href{https://www.nature.com/articles/ng.2435}{
JB Maller et al., Bayesian refinement of association signals for
 14 loci in 3 common diseases. Nature Genetics. 44, 1294–1301 (2012).}

\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1950810/}{
J Wakefield, A bayesian measure of the probability of false discovery 
in genetic epidemiology studies. American Journal of Human Genetics. 
81, 208–227 (2007).}
}
\usage{
ABF(
  dat,
  credset_thresh = 0.95,
  compute_n = "ldsc",
  sdY = NULL,
  case_control = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Fine-mapping results data.}

\item{compute_n}{How to compute per-SNP sample size (new column "N").
\itemize{
\item{\code{0}: }{N will not be computed.}
\item{\code{>0}: }{If any number >0 is provided,
that value will be set as N for every row.
\strong{Note}: Computing N this way is incorrect and should be avoided
if at all possible.}
\item{\code{"sum"}: }{N will be computed as:
cases (N_CAS) + controls (N_CON), so long as both columns are present}.
\item{\code{"ldsc"}: }{N will be computed as effective sample size:
Neff =(N_CAS+N_CON)*(N_CAS/(N_CAS+N_CON)) / mean((N_CAS/(N_CAS+N_CON))(N_CAS+N_CON)==max(N_CAS+N_CON))}.
\item{\code{"giant"}: }{N will be computed as effective sample size:
Neff = 2 / (1/N_CAS + 1/N_CON)}.
\item{\code{"metal"}: }{N will be computed as effective sample size:
Neff = 4 / (1/N_CAS + 1/N_CON)}.
}}

\item{verbose}{Print messages.}
}
\description{
Conduct statistical fine-mapping with Approximate Bayes Factor (ABF)
via \link[coloc]{finemap.abf}.
}
\examples{
dat <- echodata::LRRK2
dat2 <- echofinemap::ABF(dat=dat) 
}
