% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_required_cols.R
\name{check_required_cols}
\alias{check_required_cols}
\title{Check for necessary columns}
\usage{
check_required_cols(
  dat,
  finemap_methods = NULL,
  case_control = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{\link[data.table]{data.table} containing SNP-level 
data to fine-map.}

\item{finemap_methods}{Fine-mapping methods to run. 
See \link[echofinemap]{lfm} for a list of all fine-mapping methods currently
available.}

\item{case_control}{Whether the summary statistics come from a case-control
study (e.g. a GWAS of having Alzheimer's Disease or not) (\code{TRUE})
or a quantitative study (e.g. a GWAS of height, or an eQTL) (\code{FALSE}).}

\item{verbose}{Print messages.}
}
\description{
Check whether the input data (\code{dat}) 
has the minimum columns required
to run each fine-mapping method, as well as suggested columns.
}
\examples{
dat <- echodata::BST1
finemap_methods <- check_required_cols(dat=dat)
}
