% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_method_path.R
\name{create_method_path}
\alias{create_method_path}
\title{Create method path}
\usage{
create_method_path(
  locus_dir,
  finemap_method,
  LD_reference = NULL,
  create_dir = TRUE,
  compress = FALSE
)
}
\description{
Create a fine-mapping method-specific path.
}
\seealso{
Other finemapping functions: 
\code{\link{finemap_handler}()},
\code{\link{multi_finemap}()}
}
\concept{finemapping functions}
