% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_method_path.R
\name{create_method_path}
\alias{create_method_path}
\title{Create method path}
\usage{
create_method_path(
  locus_dir,
  finemap_method,
  prefix = "multi-finemap.tsv",
  include_astrices = FALSE,
  LD_reference = NULL,
  create_dir = TRUE,
  compress = FALSE
)
}
\description{
Create a fine-mapping method-specific path.
}
\examples{
locus_dir <- echodata::locus_dir
path <- echofinemap::create_method_path(locus_dir = locus_dir, 
                                        finemap_method = "SUSIE")
}
\seealso{
Other finemapping functions: 
\code{\link{multifinemap_handler}()},
\code{\link{multifinemap}()}
}
\concept{finemapping functions}
