% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINEMAP.R
\name{FINEMAP}
\alias{FINEMAP}
\title{Fine-map locus with \code{FINEMAP}}
\source{
\url{http://www.christianbenner.com}
}
\usage{
FINEMAP(
  dat,
  locus_dir = tempdir(),
  LD_matrix,
  FINEMAP_path = NULL,
  compute_n = "ldsc",
  n_causal = 5,
  model = "sss",
  remove_tmps = FALSE,
  force_new = TRUE,
  credset_thresh = 0.95,
  finemap_version = package_version("1.4.1"),
  prior_k = NULL,
  rescale_priors = TRUE,
  args_list = list(),
  fillNA = 0,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Fine-mapping results data.}

\item{FINEMAP_path}{Path to a custom FINEMAP executable to use
instead of the ones included in \pkg{echolocatoR}.
Users can also simply supply "finemap" if this command is linked to
 the executable.}

\item{compute_n}{How to compute per-SNP sample size (new column "N").
\itemize{
\item{\code{0}: }{N will not be computed.}
\item{\code{>0}: }{If any number >0 is provided,
that value will be set as N for every row.
\strong{Note}: Computing N this way is incorrect and should be avoided
if at all possible.}
\item{\code{"sum"}: }{N will be computed as:
cases (N_CAS) + controls (N_CON), so long as both columns are present}.
\item{\code{"effective"}: }{N will be computed as effective sample size:
cases (N_CAS) + controls (N_CON), so long as both columns are present}.
}}

\item{model}{"cond" for stepwise conditional search, 
"sss" for stochastic shotgun search.}

\item{force_new}{If saved results already exist in the given
\code{locus_dir}, skip re-running FINEMAP and use them 
(default: \code{force_new}). 
Set \code{TRUE} to ignore these files and re-run FINEMAP.}

\item{finemap_version}{Which FINEMAP version to use (specify as a string).}

\item{args_list}{A named list of additional arguments to pass to FINEMAP
(e.g.: args_list = list("--n-iterations"=5000,"--sss"="")).
Alternatively, can supply a string instead 
(e.g.: args_list = "--n-iterations 5000 --sss").}

\item{fillNA}{Fill CS/PP values without fine-mapping results 
(i.e. \code{NA}) with some default value (e.g. 0).}

\item{nThread}{Number of threads to parallelise across.
Passed to \code{"--n-threads"} in FINEMAP.}

\item{verbose}{Print messages.}
}
\description{
The stepwise conditional search starts with a causal configuration 
containing the SNP with the lowest P-value alone and then iteratively 
adds to the causal configuration the SNP given the highest 
posterior model probability until no further SNP yields
a higher posterior model probability.
}
\examples{
locus_dir <- file.path(tempdir(),echodata::locus_dir)
dat <- echodata::BST1;
LD_matrix <- echodata::BST1_LD_matrix

dir.create(file.path(locus_dir,"FINEMAP"),
 showWarnings = FALSE, recursive = TRUE)
out <- echoLD::subset_common_snps(LD_matrix, dat)
LD_matrix <- out$LD
dat <- out$DT

dat2 <- echofinemap::FINEMAP(dat=dat,
                             locus_dir=locus_dir,
                             LD_matrix=LD_matrix)
}
\seealso{
Other FINEMAP: 
\code{\link{FINEMAP_construct_data}()},
\code{\link{FINEMAP_construct_master}()},
\code{\link{FINEMAP_find_executable}()},
\code{\link{FINEMAP_process_results}()}
}
\concept{FINEMAP}
