% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(
  fname,
  repo = "RajLabMSSM/echofinemap",
  storage_dir = tempdir(),
  overwrite = FALSE
)
}
\description{
Get data via \pkg{piggyback}.
}
\keyword{internal}
