% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_finemap_methods.R
\name{lfm}
\alias{lfm}
\title{List fine-mapping methods}
\usage{
lfm(finemap_methods = NULL, verbose = TRUE)
}
\description{
List all fine-mapping methods currently available in \pkg{echofinemap}.
}
\examples{
finemap_methods <- echofinemap::list_finemap_methods()
}
