% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifinemap.R
\name{multifinemap}
\alias{multifinemap}
\title{Multi-fine-map}
\usage{
multifinemap(
  dat,
  locus_dir,
  fullSS_path,
  finemap_methods = c("ABF", "SUSIE", "FINEMAP"),
  finemap_args = NULL,
  dataset_type = "GWAS",
  force_new_finemap = TRUE,
  LD_reference = NULL,
  LD_matrix = NULL,
  n_causal = 5,
  sample_size = NULL,
  conditioned_snps = NULL,
  PAINTOR_QTL_datasets = NULL,
  PP_threshold = 0.95,
  consensus_threshold = 2,
  case_control = TRUE,
  conda_env = "echoR",
  nThread = 1,
  seed = 2022,
  verbose = TRUE
)
}
\description{
Handle fine-mapping across multiple tools.
}
\examples{
dat <- echodata::BST1
LD_matrix <- echodata::BST1_LD_matrix
locus_dir <- file.path(tempdir(),echodata::locus_dir)
fullSS_path <- echodata::example_fullSS(dataset = "Nalls2019")
dat2 <- echofinemap::multifinemap(dat = dat,
                                 locus_dir = locus_dir,
                                 LD_matrix = LD_matrix,
                                 fullSS_path = fullSS_path,
                                 finemap_methods = "SUSIE")
}
\seealso{
Other finemapping functions: 
\code{\link{create_method_path}()},
\code{\link{multifinemap_handler}()}
}
\concept{finemapping functions}
