% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifinemap_handler.R
\name{multifinemap_handler}
\alias{multifinemap_handler}
\title{Fine-map using multiple fine-mapping tools}
\usage{
multifinemap_handler(
  dat,
  locus_dir,
  fullSS_path = NULL,
  finemap_methods,
  finemap_args = NULL,
  dataset_type,
  force_new_finemap = FALSE,
  LD_matrix = NULL,
  n_causal = 5,
  compute_n = "ldsc",
  conditioned_snps = NULL,
  PAINTOR_QTL_datasets = NULL,
  PP_threshold = 0.95,
  case_control = TRUE,
  verbose = TRUE,
  nThread = 1,
  conda_env = "echoR_mini"
)
}
\arguments{
\item{dat}{Fine-mapping results data.}

\item{compute_n}{How to compute per-SNP sample size (new column "N").
\itemize{
\item{\code{0}: }{N will not be computed.}
\item{\code{>0}: }{If any number >0 is provided,
that value will be set as N for every row.
\strong{Note}: Computing N this way is incorrect and should be avoided
if at all possible.}
\item{\code{"sum"}: }{N will be computed as:
cases (N_CAS) + controls (N_CON), so long as both columns are present}.
\item{\code{"effective"}: }{N will be computed as effective sample size:
cases (N_CAS) + controls (N_CON), so long as both columns are present}.
}}

\item{verbose}{Print messages.}
}
\description{
Fine-mapping will be repeated on the same locus using each of the tools 
in \code{finemap_methods}.
Then, all results will be merged into the locus-specific multi-finemap file,
along with the original per-SNP GWAS/QTL summary statistics.
Each tools will have the following columns:
\describe{
 \item{<tool>.PP}{The posterior probability (PP) of a SNP being causal for
  the trait. Though this is a generalization and the exact meaning of PP
   will differ by tools (e.g. Posterior Inclusion Probability for SUSIE).}
 \item{<tool>.CS}{Which credible set the SNP is part of (within a locus). 
 If \code{=0}, then the SNP was not part of any credible set. 
 Some tools only produce one credible set per locus.}
}
}
\seealso{
Other finemapping functions: 
\code{\link{create_method_path}()},
\code{\link{multifinemap_handler_method}()},
\code{\link{multifinemap}()}
}
\concept{finemapping functions}
\keyword{internal}
