% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_download_annotations.R
\name{PAINTOR_download_annotations}
\alias{PAINTOR_download_annotations}
\title{Download annotations for PAINTOR}
\usage{
PAINTOR_download_annotations(
  query_dat,
  locus_dir,
  PT_results_path,
  annot_xgr = NULL,
  annot_roadmap = NA,
  chrom_states = NULL,
  no_annotations = FALSE,
  conda_env = "echoR_mini",
  nThread = 1,
  verbose = TRUE
)
}
\description{
Download annotations for PAINTOR
}
\keyword{internal}
