% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_install.R
\name{PAINTOR_install}
\alias{PAINTOR_install}
\title{Install PAINTOR via command line}
\source{
\href{https://github.com/gkichaev/PAINTOR_V3.0/issues/49}{
Installation issues with gcc versions}
\href{https://formulae.brew.sh/formula/gcc}{
Only certain subversions of gcc are available via brew}
}
\usage{
PAINTOR_install(paintor_path = NULL, force_reinstall = FALSE, verbose = TRUE)
}
\description{
Currently there is no R or conda distribution of PAINTOR_
}
\keyword{internal}
