% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_list_annotations.R
\name{PAINTOR_list_annotations}
\alias{PAINTOR_list_annotations}
\title{List available PAINTOR annotations}
\usage{
PAINTOR_list_annotations(
  annot_dir = system.file("tools", "PAINTOR_V3.0", "SampleData", package = "echofinemap"),
  pattern = "\\\\.annotations",
  recursive = TRUE,
  verbose = TRUE
)
}
\description{
List available PAINTOR annotations
}
\keyword{internal}
