% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_process_results.R
\name{PAINTOR_process_results}
\alias{PAINTOR_process_results}
\title{Process PAINTOR results}
\usage{
PAINTOR_process_results(PT_results_path, locus_name = "Locus1")
}
\description{
@keywords internal
}
