% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_run.R
\name{PAINTOR_run}
\alias{PAINTOR_run}
\title{Run PAINTOR}
\usage{
PAINTOR_run(
  paintor_path = NULL,
  PT_results_path,
  inputFile_path,
  ld_paths,
  zscore_cols,
  max_causal,
  Gname = "Enrichment.Estimates.txt",
  RESname = "results.txt",
  ANname = "annotations.txt",
  method = c("mcmc", "enumerate"),
  seed = 2022,
  auto_restart = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{paintor_path}{[Optional] Path to PAINTOR executable. 
Will be automatically installed if set to \code{NULL} (default).}

\item{max_causal}{Maximum number of causal SNPs.}

\item{method}{\code{"enumerate"} is actually faster 
when \code{max_causal} is small (<3), 
but far larger \code{max_causal} use \code{"mcmc"}.}

\item{seed}{Set the random seed for reproducible results.}

\item{auto_restart}{Automatically rerun PAINTOR if it fails the first time.}

\item{verbose}{Print messages.}
}
\description{
Run PAINTOR executable.
}
\keyword{internal}
