% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_run.R
\name{PAINTOR_run}
\alias{PAINTOR_run}
\title{Run PAINTOR}
\usage{
PAINTOR_run(
  paintor_path = NULL,
  PT_results_path,
  .LD_file.paths,
  n_datasets,
  method = "enumerate",
  n_causal = 2
)
}
\description{
@keywords internal
}
