% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_install.R
\name{POLYFUN_install}
\alias{POLYFUN_install}
\title{Install PolyFun}
\usage{
POLYFUN_install(force_new = FALSE, verbose = TRUE)
}
\arguments{
\item{force_new}{Force a new installation 
even when PolyFun is already installed.}

\item{verbose}{Print messages.}
}
\value{
Null.
}
\description{
Install PolyFun (if not already installed) from the GitHub submodule.
}
