% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_munge_summ_stats.R
\name{POLYFUN_munge_summ_stats}
\alias{POLYFUN_munge_summ_stats}
\title{Munge summary stats}
\usage{
POLYFUN_munge_summ_stats(
  polyfun = NULL,
  fullSS_path,
  locus_dir,
  sample_size = NULL,
  min_INFO = 0,
  min_MAF = 0.001,
  force_new_munge = FALSE,
  conda_env = "echoR"
)
}
\description{
Munge summary stats
}
\examples{
\dontrun{
data("genome_wide_dir");
fullSS_path <- example_fullSS(fullSS_path="~/Desktop/Nalls23andMe_2019.fullSS_subset.tsv")
munged_path <- POLYFUN_munge_summ_stats(fullSS_path=fullSS_path, locus_dir=genome_wide_dir, force_new_munge=TRUE)
}
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_SUSIE}()},
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_polyfun_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_get_precomputed_priors}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()},
\code{\link{read_parquet}()}
}
\concept{polyfun}
\keyword{internal}
