% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_munge_summ_stats.R
\name{POLYFUN_munge_summ_stats}
\alias{POLYFUN_munge_summ_stats}
\title{PolyFun: Munge summary stats}
\usage{
POLYFUN_munge_summ_stats(
  polyfun = NULL,
  fullSS_path,
  locus_dir = tempdir(),
  sample_size = NULL,
  min_INFO = 0,
  min_MAF = 0.001,
  chi2_cutoff = 30,
  keep_hla = FALSE,
  no_neff = FALSE,
  force_new_munge = FALSE,
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\description{
Munge summary statistics using the PolyFun implementation of the LDSSC
 munge sum stats python script (\code{munge_polyfun_sumstats.py}).
}
\examples{
\dontrun{
fullSS_path <- echodata::example_fullSS()
#### Remove duplicates ####
dat <- data.table::fread(fullSS_path)
munged_path <- POLYFUN_munge_summ_stats(fullSS_path=fullSS_path)
}
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_import_priors}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()},
\code{\link{POLYFUN}()}
}
\concept{polyfun}
\keyword{internal}
