% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN_SUSIE.R
\name{POLYFUN_SUSIE}
\alias{POLYFUN_SUSIE}
\title{Run PolyFun+SUSIE fine-mapping pipeline}
\source{
\url{https://www.biorxiv.org/content/10.1101/807792v3}
}
\usage{
POLYFUN_SUSIE(
  locus_dir,
  polyfun = NULL,
  dat = NULL,
  LD_matrix = NULL,
  polyfun_approach = "non-parametric",
  dataset_type = "GWAS",
  max_causal = 5,
  sample_size = NULL,
  PP_threshold = 0.95,
  conda_env = "echoR"
)
}
\description{
Uses echolocatoR wrapper for SUSIE instead of the \code{POLYFUN_finemapper}.
function which uses a python script provided with PolyFun.
}
\examples{
locus_dir <- file.path(tempdir(),echodata::locus_dir)

dat <- echodata::BST1
dat2 <- echofinemap::POLYFUN_SUSIE(locus_dir=)
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN_compute_priors}()},
\code{\link{POLYFUN_download_ref_files}()},
\code{\link{POLYFUN_find_polyfun_folder}()},
\code{\link{POLYFUN_finemapper}()},
\code{\link{POLYFUN_functional_enrichment}()},
\code{\link{POLYFUN_gather_annotations}()},
\code{\link{POLYFUN_gather_ldscores}()},
\code{\link{POLYFUN_get_precomputed_priors}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN_h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN_h2_enrichment}()},
\code{\link{POLYFUN_help}()},
\code{\link{POLYFUN_initialize}()},
\code{\link{POLYFUN_ldsc_annot_enrichment}()},
\code{\link{POLYFUN_munge_summ_stats}()},
\code{\link{POLYFUN_plot}()},
\code{\link{POLYFUN_prepare_snp_input}()},
\code{\link{POLYFUN_run_ldsc}()},
\code{\link{read_parquet}()}
}
\concept{polyfun}
