% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUSIE.R
\name{SUSIE}
\alias{SUSIE}
\title{Fine-map with SUSIE}
\source{
\href{https://stephenslab.github.io/susieR/}{GitHub}
\href{https://rss.onlinelibrary.wiley.com/doi/full/10.1111/rssb.12388}{
Publication}
}
\usage{
SUSIE(
  dat,
  LD_matrix,
  case_control = TRUE,
  max_causal = 5,
  compute_n = "ldsc",
  priors_col = NULL,
  rescale_priors = TRUE,
  credset_thresh = 0.95,
  scaled_prior_variance = 0.001,
  estimate_residual_variance = FALSE,
  estimate_prior_variance = TRUE,
  residual_variance = NULL,
  max_iter = 100,
  estimate_prior_method = "optim",
  var_y = NULL,
  plot_track_fit = FALSE,
  return_all_CS = TRUE,
  file_prefix = file.path(tempdir(), "SUSIE"),
  verbose = TRUE
)
}
\arguments{
\item{dat}{Fine-mapping results data.}

\item{LD_matrix}{Linkage Disequilibrium (LD) matrix to use for fine-mapping.}

\item{case_control}{Whether the summary statistics come from a case-control
study (e.g. a GWAS of having Alzheimer's Disease or not) (\code{TRUE})
or a quantitative study (e.g. a GWAS of height, or an eQTL) (\code{FALSE}).}

\item{max_causal}{The maximum number of non-zero effects 
(and thus causal variants).}

\item{compute_n}{How to compute per-SNP sample size (new column "N").\cr
If the column "N" is already present in \code{dat}, this column 
will be used to extract per-SNP sample sizes
and the argument \code{compute_n} will be ignored.\cr
If the column "N" is \emph{not} present in \code{dat}, one of the following 
options can be supplied to \code{compute_n}: 
\itemize{
\item{\code{0}: }{N will not be computed.}
\item{\code{>0}: }{If any number >0 is provided,
that value will be set as N for every row.
**Note**: Computing N this way is incorrect and should be avoided
if at all possible.}
\item{\code{"sum"}: }{N will be computed as:
cases (N_CAS) + controls (N_CON), so long as both columns are present}.
\item{\code{"ldsc"}: }{N will be computed as effective sample size:
Neff =(N_CAS+N_CON)*(N_CAS/(N_CAS+N_CON)) / mean((N_CAS/(N_CAS+N_CON))(N_CAS+N_CON)==max(N_CAS+N_CON))}.
\item{\code{"giant"}: }{N will be computed as effective sample size:
Neff = 2 / (1/N_CAS + 1/N_CON)}.
\item{\code{"metal"}: }{N will be computed as effective sample size:
Neff = 4 / (1/N_CAS + 1/N_CON)}.
}}

\item{priors_col}{[Optional] Name of the a column in 
\code{dat} to extract SNP-wise prior probabilities from.}

\item{rescale_priors}{If prior probabilities are supplied,
rescale them from 0-1 (i.e. \code{rescaled_priors = priors / sum(priors)}).}

\item{credset_thresh}{The minimum mean Posterior Probability
(across all fine-mapping methods used) of SNPs to be included in 
the "mean.CS" column.}

\item{scaled_prior_variance}{The prior variance, divided by
\code{var(y)} (or by \code{(1/(n-1))yty} for
\code{susie_suff_stat}); that is, the prior variance of each
non-zero element of b is \code{var(y) * scaled_prior_variance}. The
value provided should be either a scalar or a vector of length
\code{L}. If \code{estimate_prior_variance = TRUE}, this provides
initial estimates of the prior variances.}

\item{estimate_residual_variance}{If
\code{estimate_residual_variance = TRUE}, the residual variance is
estimated, using \code{residual_variance} as an initial value. If
\code{estimate_residual_variance = FALSE}, the residual variance is
fixed to the value supplied by \code{residual_variance}.}

\item{estimate_prior_variance}{If \code{estimate_prior_variance =
TRUE}, the prior variance is estimated (this is a separate
parameter for each of the L effects). If provided,
\code{scaled_prior_variance} is then used as an initial value for
the optimization. When \code{estimate_prior_variance = FALSE}, the
prior variance for each of the L effects is determined by the
value supplied to \code{scaled_prior_variance}.}

\item{residual_variance}{Variance of the residual. If
\code{estimate_residual_variance = TRUE}, this value provides the
initial estimate of the residual variance. By default, it is set to
\code{var(y)} in \code{susie} and \code{(1/(n-1))yty} in
\code{susie_suff_stat}.}

\item{max_iter}{Maximum number of IBSS iterations to perform.}

\item{estimate_prior_method}{The method used for estimating prior
variance. When \code{estimate_prior_method = "simple"} is used, the
likelihood at the specified prior variance is compared to the
likelihood at a variance of zero, and the setting with the larger
likelihood is retained.}

\item{var_y}{[Optional] User-supplied phenotypic variance value(s). 
Can be one of the following:
\itemize{
\item{\code{NULL}: }{Variance will be inferred automatically by SUSIE.}
\item{Numeric vector: }{Variance will be computed directly from vector.}
\item{Character string: }{The name of a column in \code{dat}
 to extract a numeric vector from to compute variance.}
\item{"case_control"}{Variance will be inferred from the proportion 
of cases/controls in the study. 
Only works when both "N_cases" and "N_controls" are columns in \code{dat}.}
}}

\item{plot_track_fit}{Record each iteration and make a GIF of the
fine-mapping algorithm learning the causal variants.
\strong{WARNING!:} Making this plot can take a long time if there's
 many iterations.}

\item{return_all_CS}{If >1 Credible Set is identified, 
return them all (\code{TRUE}), or just the first (\code{FALSE}).}

\item{file_prefix}{Prefix to path of output plot file. If not
specified, the plot, or plots, will be saved to a temporary
directory generated using \code{\link{tempdir}}.}

\item{verbose}{Print messages.}
}
\description{
Sum of Single Effects (SuSiE): Iterative Bayesian Step-wise Selection.
}
\details{
\strong{Notes on convergence:}
\pkg{susieR} will often give the warning: 
\code{IBSS algorithm did not converge in 100 iterations!}.
This means the results might not necessarily be reliable.
There's several things you can try to avoid this:
\itemize{
\item{Make sure \code{susieR} is up-to-date: 
\code{ devtools::install_github("stephenslab/susieR@0.9.0")}}
\item{Increase \code{max_causal} (e.g. 5 => 10).}
\item{Increase \code{max_iter} (e.g. 100 => 1000), 
though this will take longer.}
\item{Decrease the locus window size, which will also speed up
 the algorithm but potentially miss causal variants far from the lead SNP.}
}
Changing \code{estimate_prior_method} does not seem to affect 
convergence warnings.

\strong{Notes on variance:}
\href{https://github.com/stephenslab/susieR/issues/90}{GitHub Issue}
If \code{estimate_residual_variance=TRUE} \emph{without} 
providing \code{var_y} \emph{and} \code{L>1}, \pkg{susieR} will throw error:
\code{Estimating residual variance failed: the estimated value is negative}
Running \pkg{susieR} with \code{var_y = var(b)} provides \emph{exactly}
 the same results.
}
\examples{
dat <- echodata::BST1
LD_matrix <- echodata::BST1_LD_matrix
dat2 <- echofinemap::SUSIE(dat=dat, LD_matrix=LD_matrix)
}
